+++ frontend/app/components/layout/form/list-form.component.ts
@@ -20,67 +20,91 @@ export class ModulesListFormComponent extends ModulesLayoutComponent implements 
     this.bPostComputeLayout = true;
   }
 
+  // la liste
   items: any[] = [];
+
   // pour pouvoir filter en local
   itemsSave: any[] = [];
+
+  // nombre d'items
   nbItems;
 
   /** fonction de comparaison pour mat-select */
-  compareFn = (a, b) => a > b;
+  compareFn = (a, b) => a == b;
 
+  // pour les recheches
   search = '';
   searchSubject: Subject<string>;
   searchSubscription = null;
+
+  // chargement en cours
   isLoading = false;
 
+  // les options
   listFormOptions;
 
   @Input() formControl: AbstractControl = null;
 
-  initSearch() {
-    this.searchSubject = this.searchSubject || new Subject();
-    this.searchSubscription && this.searchSubscription.unsubscribe();
-    this.searchSubscription = this.searchSubject
-      .pipe(distinctUntilChanged(), debounceTime(400 * this.listFormOptions.reload_on_search))
-      .subscribe((search) => {
-        this.processSearchChanged(search);
-      });
-  }
-
-  getCompareFn = (return_object, value_field_name) => (a, b) => {
-    return a && b && return_object ? a[value_field_name] == b[value_field_name] : a == b;
-  };
-  /** initialisation du composant select */
-  // postProcessLayout(): void {
-
-  // }
-
-  processListFormConfig() {
-    this.listFormOptions = utils.copy(this.computedLayout);
-    this.listFormOptions.module_code = this.context.module_code;
-    this.initSearch();
-    this.isLoading = true;
-    this._listFormService
-      .initListForm(this.listFormOptions, this.formControl)
-      .subscribe((infos) => {
-        this.compareFn = this.getCompareFn(
-          this.listFormOptions.return_object,
-          this.listFormOptions.value_field_name
-        );
-
-        this.items = infos.items;
-        this.nbItems = infos.nbItems;
-        this.itemsSave = utils.copy(infos.items);
-        this.isLoading = false;
-        // mettre à jour l'affichage des données
-      });
-  }
-
-  /** */
   postComputeLayout(dataChanged, layoutChanged): void {
     if (layoutChanged) {
       this.processListFormConfig();
     }
+  }
+
+  /** process de la configuration du composant */
+  processListFormConfig() {
+    // options de le layout
+    this.listFormOptions = utils.copy(this.computedLayout);
+
+    // module_code depuis le context
+    this.listFormOptions.module_code = this.context.module_code;
+
+    // initialisation de la recherche ??
+    this.initSearch();
+
+    // chargement des données en cours
+    this.isLoading = true;
+
+    // initialisation
+    this._listFormService
+      .initListForm(this.listFormOptions, this.formControl)
+      .subscribe((infos) => {
+        // fonction de comparaison
+        this.compareFn = this._listFormService.compareFn(
+          this.listFormOptions.return_object,
+          this.listFormOptions.value_field_name
+        );
+
+        // la liste
+        this.items = infos.items;
+
+        // le nombre d'items
+        this.nbItems = infos.nbItems;
+
+        // pour la recherche en local
+        this.itemsSave = utils.copy(infos.items);
+
+        // fin du chargement
+        this.isLoading = false;
+      });
+  }
+
+  // pour la recherche
+  initSearch() {
+    // sujet
+    this.searchSubject = this.searchSubject || new Subject();
+
+    // unsubscribe si besoin
+    this.searchSubscription && this.searchSubscription.unsubscribe();
+
+    // debounce time dans le cas autocomplete (reload on search)
+    const dt = this.listFormOptions.reload_on_search ? 400 : 0;
+
+    this.searchSubscription = this.searchSubject
+      .pipe(distinctUntilChanged(), debounceTime(dt))
+      .subscribe((search) => {
+        this.processSearchChanged(search);
+      });
   }
 
   /** action lorsque le champs de recherche change */
@@ -90,42 +114,80 @@ export class ModulesListFormComponent extends ModulesLayoutComponent implements 
 
   /** processus de recherche dans les items */
   processSearchChanged(search) {
+    // recherche "server side"
+    // - appel de l'api avec des parametre de filtre
     if (this.listFormOptions.reload_on_search) {
-      // server side search ??
-      this.listFormOptions.search = search;
-      this.isLoading = true;
-      this._listFormService
-        .getSelectList(this.listFormOptions, this.formControl.value)
-        .subscribe((infos) => {
-          this.items = infos.items;
-          this.nbItems = infos.nbItems;
-          this.isLoading = false;
-        });
+      this.remoteSearch(search);
+      // recherche locale
+      // - filtrage de la liste selon la valeur de search
     } else {
-      // local search
       this.items = this.localSearch(search);
     }
   }
 
+  /** Server Side Search */
+  remoteSearch(search) {
+    // option de recherhe
+    this.listFormOptions.search = search;
+
+    // chargement en cours
+    this.isLoading = true;
+
+    // appel de l'api
+    // on prend en compte les valeurs courrantes
+    this._listFormService
+      .getSelectList(this.listFormOptions, this.formControl.value)
+      .subscribe((infos) => {
+        // traitement des résultats
+        this.items = infos.items;
+        this.nbItems = infos.nbItems;
+        this.isLoading = false;
+      });
+  }
+
   /** Local search */
   localSearch(search) {
+    // si search est null ou ''
+    // -> on renvoie toute la liste (itemsSave)
     if ([null, undefined, ''].includes(search)) {
       return this.itemsSave;
     }
+
+    // on recherche sans prendre en compte la casse ou les accents
+    // -> on tranforme search
     const searchLowerUnaccent = utils.lowerUnaccent(search);
+
+    // filtrage selon searchLowerUnaccent
     return this.itemsSave.filter((item) => {
-      const label = utils.lowerUnaccent(item[this.listFormOptions.label_field_name]);
+      let label = item[this.listFormOptions.label_field_name];
+
+      // on ne tient pas compte de la casse et des accents
+      label = utils.lowerUnaccent(item[this.listFormOptions.label_field_name]);
+
+      // si label contient searchLowerUnaccent
+      // -> il est dans la liste
       return label.includes(searchLowerUnaccent);
     });
   }
 
-  /** remise à zéro de la selection */
+  /** remise à zéro de la selection
+   * - action pour le boutton 'clear'
+   */
   clearSelection(event) {
+    // ne pas propager l'action
+    // sinon on a des comportements bizzare (validation du formulaire ou autre)
     event.stopPropagation();
-    this.formControl.patchValue(this.listFormOptions.multiple ? [] : null);
+
+    // prise en compte de l'option multiple
+    const value = this.listFormOptions.multiple ? [] : null;
+
+    // on donne la valeur de value au formulaire
+    this.formControl.patchValue(value);
   }
 
-  /** Quand le formulaire est modifié */
+  /** Action quand l'input du select est modifié
+   * si une fonction change est définie dans les options
+   */
   inputChange() {
     this.listFormOptions.change && this.listFormOptions.change();
   }